+++ src/jws.rs
@@ -1,8 +1,8 @@
 //! Jws Implementation
 
-use crate::crypto::{Jwk, JwsCompact};
+use crate::compact::{Jwk, JwsCompact, JwsInner};
 #[cfg(feature = "openssl")]
-use crate::crypto::{JwsInner, JwsSigner, JwsValidator};
+use crate::crypto::{JwsSigner, JwsValidator};
 #[cfg(feature = "openssl")]
 use openssl::x509;
 #[cfg(feature = "openssl")]
